fix: Sleep Schedule Modal is functional
