+++ client/components/calendar-add.js
@@ -1,13 +1,12 @@
 import React, {Component} from 'react'
-import ReactModal from 'react-modal'
+import Modal from 'styled-react-modal'
 import {connect} from 'react-redux'
-import {generateSleepEvents, loadEvents} from '../store/calendar'
+import {loadEvents} from '../store/calendar'
+import {numbersToEvents} from './sleepAlgorithm'
 
 export const gapi = window.gapi
 export const CLIENT_ID = process.env.GOOGLE_CALENDAR_CLIENT_ID
 export const API_KEY = process.env.GOOGLE_CALENDAR_API_KEY
-console.log('process.env keys: ' + Object.keys(process.env))
-
 // Array of API discovery doc URLs for APIs used by the quickstart
 export const DISCOVERY_DOCS = [
   'https://www.googleapis.com/discovery/v1/apis/calendar/v3/rest',
@@ -15,25 +14,40 @@ export const DISCOVERY_DOCS = [
 // Authorization scopes required by the API; multiple scopes can be included, separated by spaces.
 export const SCOPES = 'https://www.googleapis.com/auth/calendar.events'
 
+const StyledModal = Modal.styled`
+  width: 400px;
+  height: 500px;
+  display: flex;
+  color: black;
+  padding: 40px;
+  flex-flow: column wrap;
+  align-items: center;
+  justify-content: space-around;
+  background-color: white;
+  opacity: 85%;
+`
+
 class AddToCalendar extends Component {
   constructor() {
     super()
     this.state = {
       showModal: false,
+      sleepEvents: [],
     }
     this.handleClick = this.handleClick.bind(this)
     this.openModal = this.openModal.bind(this)
     this.closeModal = this.closeModal.bind(this)
   }
-  componentDidMount() {
-    this.props.loadEvents()
-    ReactModal.setAppElement('body')
+  async componentDidMount() {
+    await this.props.loadEvents()
   }
 
   openModal() {
     this.setState({showModal: true})
-    this.props.generateSleepEvents(this.props.events)
+    const sleepEvents = numbersToEvents(this.props.events)
+    this.setState({sleepEvents})
   }
+
   closeModal() {
     this.setState({showModal: false})
   }
@@ -47,7 +61,8 @@ class AddToCalendar extends Component {
   }
 
   render() {
-    const {events} = this.props
+    const {sleepEvents} = this.state
+
     return (
       <div>
         <input
@@ -58,44 +73,54 @@ class AddToCalendar extends Component {
           onClick={this.openModal}
         />
 
-        <ReactModal className="modal-content" isOpen={this.state.showModal}>
-          {events.map((sleepEvent, i) => {
-            return (
-              <div key={i}>
-                {sleepEvent.summary === 'nap' ? (
-                  <div>
-                    <p> Napping:</p>
-                    From {sleepEvent.start.dateTime} to{' '}
-                    {sleepEvent.end.dateTime}
-                    <button
-                      type="button"
-                      onClick={() => this.handleClick(sleepEvent)}
-                    >
-                      Add to Calendar
-                    </button>
-                  </div>
-                ) : (
-                  <div>
-                    <p>Full 8-hour Sleep Cycle:</p>
-                    From {sleepEvent.start.dateTime} to{' '}
-                    {sleepEvent.end.dateTime}
-                    <button
-                      type="button"
-                      onClick={() => this.handleClick(sleepEvent)}
-                    >
-                      Add to Calendar
-                    </button>
-                  </div>
-                )}
-              </div>
-            )
-          })}
+        <StyledModal
+          isOpen={this.state.showModal}
+          allowScroll="true"
+          onBackgroundClick={this.closeModal}
+          onEscapeKeydown={this.closeModal}
+        >
+          {sleepEvents &&
+            sleepEvents.map((sleepEvent, i) => {
+              return (
+                <div key={i}>
+                  {sleepEvent.summary === 'nap' ? (
+                    <div>
+                      <p>Napping:</p>
+                      From {sleepEvent.start.dateTime} to{' '}
+                      {sleepEvent.end.dateTime}
+                      <p>
+                        <button
+                          type="button"
+                          onClick={() => this.handleClick(sleepEvent)}
+                        >
+                          Add to Calendar
+                        </button>
+                      </p>
+                    </div>
+                  ) : (
+                    <div>
+                      <p>Full 8-hour Sleep Cycle:</p>
+                      From {sleepEvent.start.dateTime} to{' '}
+                      {sleepEvent.end.dateTime}
+                      <p>
+                        <button
+                          type="button"
+                          onClick={() => this.handleClick(sleepEvent)}
+                        >
+                          Add to calendar
+                        </button>
+                      </p>
+                    </div>
+                  )}
+                </div>
+              )
+            })}
           <p>
             <button type="button" onClick={this.closeModal}>
               Done
             </button>
           </p>
-        </ReactModal>
+        </StyledModal>
       </div>
     )
   }
@@ -104,7 +129,6 @@ class AddToCalendar extends Component {
 const mapState = ({events}) => ({events})
 const mapDispatch = (dispatch) => {
   return {
-    generateSleepEvents: (events) => dispatch(generateSleepEvents(events)),
     loadEvents: () => dispatch(loadEvents()),
   }
 }
