+++ package.json
@@ -44,6 +44,7 @@
     "axios": "^0.19.2",
     "compression": "^1.7.4",
     "connect-session-sequelize": "^7.0.1",
+    "css-loader": "^5.0.1",
     "dotenv": "^8.2.0",
     "express": "^4.17.1",
     "express-session": "^1.17.1",
@@ -56,15 +57,20 @@
     "pg-hstore": "^2.3.3",
     "prop-types": "^15.7.2",
     "react": "^16.13.1",
+    "react-big-calendar": "^0.28.5",
     "react-dom": "^16.13.1",
-    "react-modal": "^3.11.2",
     "react-redux": "^7.2.1",
     "react-router-dom": "^5.2.0",
     "redux": "^4.0.5",
     "redux-logger": "^3.0.6",
     "redux-thunk": "^2.3.0",
     "sequelize": "^6.3.3",
-    "socket.io": "^2.3.0"
+    "socket.io": "^2.3.0",
+    "style-loader": "^2.0.0",
+    "styled-components": "^5.2.1",
+    "styled-react-modal": "^2.0.1",
+    "webpack": "^4.44.0",
+    "webpack-cli": "^3.3.12"
   },
   "devDependencies": {
     "@babel/core": "^7.10.5",
@@ -103,8 +109,6 @@
     "redux-mock-store": "^1.5.4",
     "simple-git": "^2.15.0",
     "supertest": "^4.0.2",
-    "webpack": "^4.44.0",
-    "webpack-cli": "^3.3.12",
     "yaml": "^1.10.0"
   }
 }
