+++ server/api/index.js
@@ -2,6 +2,7 @@ const router = require('express').Router()
 module.exports = router
 
 router.use('/users', require('./users'))
+// router.use('/events', require('./events'))
 
 router.use((req, res, next) => {
   const error = new Error('Not Found')
