+++ webpack.config.js
@@ -42,6 +42,9 @@ module.exports = {
         test: /\.jsx?$/,
         exclude: /node_modules/,
         loader: 'babel-loader',
+        // use: [{ loader: 'style-loader' },
+        // { loader: 'css-loader' },
+        // { loader: 'babel-loader' }]
       },
     ],
   },
